Added mouse.gui package: Basic JFrame, Canvas and Drawable Interface for the structure of double buffered anti aliased 2D drawing in Java Swing/AWT.
Note: All graphics are real just for testing purpose at the moment and will be polished and replaced with image drawing, scaling, window resize, full screen, bla bla.

Added: Pair<F,S> which is exactly just a pair of values

Modified: Level to return also a Pair<Point,Orientation> for the Mouse.

Still Missing: I added a lot of TODO entries which i need to fix, but the code is pretty much stand alone and compile/runnable so i thought it might be a perfect point for a commit.