Added Server Client communication: You can connect and the mouse will move. Added Client refresh on update (you can see changes now) Added some debug messages Added a new GameLogic Event and Listener group of classes. They act as a layer between network and logic. I really wanted to differentiate between the network Messages and events the server handles.
The game needs to update way faster (the input lag from opening and closing doors feels strange)
The communication could actually be improved by factor N (where N is the number of players) by simply sending one message with all mouse and door state changes instead of one message per door and one message per mouse.

The FileLevel from Simon is NOT yet added but simply the configuration file needs to be adapted from int to String and instead of loading the predefined level loading needs to be replaced with the FileLevel-Loading.